Extract UserInfoResolverConfig as a separate class
